+++ src/HIndent/Printer.hs
@@ -15,7 +15,6 @@ import Control.Applicative
 import Control.Monad
 import Control.Monad.State.Strict
 import Data.ByteString.Builder
-import Data.Int (Int64)
 import HIndent.Config
 
 -- | A pretty printing monad.
@@ -31,16 +30,16 @@ newtype Printer a = Printer
 
 -- | The state of the pretty printer.
 data PrintState = PrintState
-  { psIndentLevel :: !Int64
+  { psIndentLevel :: !Int
     -- ^ Current indentation level, i.e. every time there's a
     -- new-line, output this many spaces.
   , psOutput :: !Builder
     -- ^ The current output bytestring builder.
   , psNewline :: !Bool
     -- ^ Just outputted a newline?
-  , psColumn :: !Int64
+  , psColumn :: !Int
     -- ^ Current column.
-  , psLine :: !Int64
+  , psLine :: !Int
     -- ^ Current line number.
   , psConfig :: !Config
     -- ^ Configuration of max colums and indentation style.
